backend/ml_engine/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
requests
python-dotenv
tensorflow
pillow
numpy
python-multipart
beautifulsoup4
pymongo
